Manage hash values with size_t in hash tables of elements associated to output files
git-svn-id: https://forge.ipsl.fr/ioserver/svn/XIOS3/trunk@2388 d613d254-d18a-4304-b296-5125d4c4d670

+++ src/attribute.hpp
@@ -43,7 +43,7 @@ namespace xios
             virtual StdString dump(void) const = 0;
             virtual StdString dumpGraph(void) const = 0;
             virtual bool isEqual(const CAttribute& ) = 0;
-            virtual int computeHash(void) = 0;
+            virtual size_t computeHash(void) = 0;
 
             //! Returns true if and only if the attribute should be publicly exposed in the API
             virtual bool isPublic() const { return true; }
