+++ src/attribute_map.hpp
@@ -36,7 +36,7 @@ namespace xios
             void setAttributes(const xml::THashAttributes & attributes);
             void setAttributes(const CAttributeMap * const _parent, bool apply=true);
             void duplicateAttributes(const CAttributeMap* const _parent);
-            int computeGlobalAttributesHash(const vector<StdString>& excludedAttrs);
+            size_t computeGlobalAttributesHash(const vector<StdString>& excludedAttrs);
             void clearAllAttributes(void);
             StdString dumpXiosAttributes(void) const;
             StdString recordXiosAttributes(void) const;
